update: put streaming client and ws adapter in a seperate files
