+++ cmd/jaeger/internal/extension/jaegerquery/internal/jaegerai/handler.go
@@ -7,9 +7,7 @@ import (
 	"context"
 	"encoding/json"
 	"fmt"
-	"io"
 	"net/http"
-	"sync"
 	"time"
 
 	"github.com/coder/acp-go-sdk"
@@ -20,51 +18,6 @@ import (
 )
 
 const endOfTurnMarker = "__END_OF_TURN__"
-
-// WsReadWriteCloser wraps a gorilla websocket to implement io.ReadWriteCloser
-type WsReadWriteCloser struct {
-	conn *websocket.Conn
-	r    io.Reader
-}
-
-func NewWsAdapter(conn *websocket.Conn) *WsReadWriteCloser {
-	return &WsReadWriteCloser{conn: conn}
-}
-
-func (w *WsReadWriteCloser) Read(p []byte) (int, error) {
-	if w.r == nil {
-		messageType, r, err := w.conn.NextReader()
-		if err != nil {
-			return 0, err
-		}
-		if messageType != websocket.TextMessage && messageType != websocket.BinaryMessage {
-			return 0, fmt.Errorf("unexpected message type: %d", messageType)
-		}
-		w.r = r
-	}
-
-	n, err := w.r.Read(p)
-	if err == io.EOF {
-		w.r = nil
-		if n > 0 {
-			return n, nil
-		}
-		return w.Read(p)
-	}
-	return n, err
-}
-
-func (w *WsReadWriteCloser) Write(p []byte) (int, error) {
-	err := w.conn.WriteMessage(websocket.TextMessage, p)
-	if err != nil {
-		return 0, err
-	}
-	return len(p), nil
-}
-
-func (w *WsReadWriteCloser) Close() error {
-	return w.conn.Close()
-}
 
 // ChatRequest is the incoming payload
 type ChatRequest struct {
@@ -79,148 +32,6 @@ type ChatHandler struct {
 
 func NewChatHandler(logger *zap.Logger, queryService *querysvc.QueryService) *ChatHandler {
 	return &ChatHandler{Logger: logger, QueryService: queryService}
-}
-
-// streamingClient implements acp.Client to handle callbacks and streaming text
-type streamingClient struct {
-	requestCtx context.Context
-	w          http.ResponseWriter
-	flusher    http.Flusher
-	mu         sync.Mutex
-	closed     bool
-	doneCh     chan struct{}
-	doneOnce   sync.Once
-}
-
-func (c *streamingClient) signalDone() {
-	c.doneOnce.Do(func() {
-		if c.doneCh != nil {
-			close(c.doneCh)
-		}
-	})
-}
-
-func (c *streamingClient) writeAndFlush(text string) {
-	c.mu.Lock()
-	defer c.mu.Unlock()
-
-	if c.closed {
-		return
-	}
-
-	if c.requestCtx != nil {
-		select {
-		case <-c.requestCtx.Done():
-			c.closed = true
-			c.signalDone()
-			return
-		default:
-		}
-	}
-
-	defer func() {
-		if recover() != nil {
-			c.closed = true
-			c.signalDone()
-		}
-	}()
-
-	if _, err := io.WriteString(c.w, text); err != nil {
-		c.closed = true
-		c.signalDone()
-		return
-	}
-
-	c.flusher.Flush()
-}
-
-func (c *streamingClient) waitForTurnCompletion(ctx context.Context, maxWait time.Duration) {
-	if maxWait <= 0 {
-		return
-	}
-
-	maxTimer := time.NewTimer(maxWait)
-	defer maxTimer.Stop()
-
-	select {
-	case <-ctx.Done():
-		return
-	case <-maxTimer.C:
-		return
-	case <-c.doneCh:
-		return
-	}
-}
-
-func (*streamingClient) RequestPermission(_ context.Context, p acp.RequestPermissionRequest) (acp.RequestPermissionResponse, error) {
-	if len(p.Options) == 0 {
-		return acp.RequestPermissionResponse{
-			Outcome: acp.RequestPermissionOutcome{
-				Cancelled: &acp.RequestPermissionOutcomeCancelled{},
-			},
-		}, nil
-	}
-	return acp.RequestPermissionResponse{
-		Outcome: acp.RequestPermissionOutcome{
-			Selected: &acp.RequestPermissionOutcomeSelected{OptionId: p.Options[0].OptionId},
-		},
-	}, nil
-}
-
-func (c *streamingClient) SessionUpdate(_ context.Context, n acp.SessionNotification) error {
-	u := n.Update
-	if u.AgentMessageChunk != nil {
-		content := u.AgentMessageChunk.Content
-		if content.Text != nil {
-			if content.Text.Text == endOfTurnMarker {
-				c.signalDone()
-			} else {
-				c.writeAndFlush(content.Text.Text)
-			}
-		}
-	}
-	if u.ToolCall != nil {
-		c.writeAndFlush(fmt.Sprintf("\n[tool_call] %s\n", u.ToolCall.Title))
-	}
-	if u.ToolCallUpdate != nil {
-		c.writeAndFlush(fmt.Sprintf("\n[tool_result] id=%s status=%s\n", u.ToolCallUpdate.ToolCallId, valueOrUnknown(u.ToolCallUpdate.Status)))
-	}
-	return nil
-}
-
-func valueOrUnknown(v *acp.ToolCallStatus) string {
-	if v == nil {
-		return "unknown"
-	}
-	return string(*v)
-}
-
-func (*streamingClient) WriteTextFile(_ context.Context, _ acp.WriteTextFileRequest) (acp.WriteTextFileResponse, error) {
-	return acp.WriteTextFileResponse{}, nil
-}
-
-func (*streamingClient) ReadTextFile(_ context.Context, p acp.ReadTextFileRequest) (acp.ReadTextFileResponse, error) {
-	return acp.ReadTextFileResponse{Content: "unsupported path: " + p.Path}, nil
-}
-
-func (*streamingClient) CreateTerminal(_ context.Context, _ acp.CreateTerminalRequest) (acp.CreateTerminalResponse, error) {
-	return acp.CreateTerminalResponse{TerminalId: "t-1"}, nil
-}
-
-func (*streamingClient) KillTerminalCommand(_ context.Context, _ acp.KillTerminalCommandRequest) (acp.KillTerminalCommandResponse, error) {
-	return acp.KillTerminalCommandResponse{}, nil
-}
-
-func (*streamingClient) ReleaseTerminal(_ context.Context, _ acp.ReleaseTerminalRequest) (acp.ReleaseTerminalResponse, error) {
-	return acp.ReleaseTerminalResponse{}, nil
-}
-
-func (*streamingClient) TerminalOutput(_ context.Context, _ acp.TerminalOutputRequest) (acp.TerminalOutputResponse, error) {
-	return acp.TerminalOutputResponse{Output: "ok", Truncated: false}, nil
-}
-
-func (*streamingClient) WaitForTerminalExit(_ context.Context, _ acp.WaitForTerminalExitRequest) (acp.WaitForTerminalExitResponse, error) {
-	return acp.WaitForTerminalExitResponse{}, nil
 }
 
 func (h *ChatHandler) ServeHTTP(w http.ResponseWriter, r *http.Request) {
