+++ python-sidecar/main.py
@@ -1,11 +1,14 @@
 import asyncio
+import websockets
 
-from sidecar import main as sidecar_main
+from sidecar import handle_websocket
 
 
-def main():
-    asyncio.run(sidecar_main())
+async def main():
+    async with websockets.serve(handle_websocket, "localhost", 9000):
+        print("Jaeger ACP Sidecar listening on ws://localhost:9000")
+        await asyncio.Future()
 
 
 if __name__ == "__main__":
-    main()
+    asyncio.run(main())
