+++ python-sidecar/sidecar.py
@@ -288,11 +288,3 @@ async def handle_websocket(websocket):
         task.cancel()
         
     print("Websocket connection closed")
-
-async def main():
-    async with websockets.serve(handle_websocket, "localhost", 9000):
-        print("Jaeger ACP Sidecar listening on ws://localhost:9000")
-        await asyncio.Future()
-
-if __name__ == "__main__":
-    asyncio.run(main())
